Fixed many checker errors; Updated Kernel,Datastructures and Math to OpenMS type definitions; Added methods for Real/UInt to DataValue, Param, MetaInfo and MetainfoInterface; change to FFPP
git-svn-id: https://open-ms.svn.sourceforge.net/svnroot/open-ms/OpenMS@2683 6adb6e08-d915-0410-941f-83917bcadc18

+++ include/OpenMS/ANALYSIS/MAPMATCHING/BasePairwiseMapMatcher.h
@@ -133,7 +133,7 @@ namespace OpenMS
     void setNumberBuckets(UInt dim, UInt number)
     {
       number_buckets_[dim] = number;
-			param_.setValue(String("number_buckets:") + RawDataPoint2D::shortDimensionName(dim), (Int)number);
+			param_.setValue(String("number_buckets:") + RawDataPoint2D::shortDimensionName(dim), number);
     }
 
     /// Get number of buckets in dimension index
