+++ source/FILTERING/SMOOTHING/SavitzkyGolaySVDFilter.C
@@ -46,7 +46,7 @@ namespace OpenMS
   void SavitzkyGolaySVDFilter::setWindowSize(UInt frame_size)
   {
     frame_size_=frame_size;
-    param_.setValue("frame_length",(Int)frame_size_);
+    param_.setValue("frame_length",frame_size_);
     
     coeffs_.clear();
     coeffs_.resize(frame_size_*(frame_size_/2+1));
@@ -56,7 +56,7 @@ namespace OpenMS
   void SavitzkyGolaySVDFilter::setOrder(UInt order)
   {
     order_=order;
-    param_.setValue("polynomial_order",(Int)order_);
+    param_.setValue("polynomial_order",order_);
     
     computeCoeffs_();
   }
